correct parcel media command cast for both X and Y, round 2, Thanks again Ubit

+++ OpenSim/Region/ScriptEngine/Shared/Api/Implementation/LSL_Api.cs
@@ -13130,7 +13130,7 @@ namespace OpenSim.Region.ScriptEngine.Shared.Api
             {
             if (commandList.Data[i] is LSL_Integer)
 			    cmd = (LSL_Integer)commandList.Data[i]; // Yengine casting issue
-			else cmd = commandList.Data[i];
+			else cmd = (int)commandList.Data[i];
 		    ParcelMediaCommandEnum command = (ParcelMediaCommandEnum)cmd;
 		  // ParcelMediaCommandEnum command = (ParcelMediaCommandEnum)commandList.Data[i];
                 switch (command)
